+++ src/createWebpackConfig.js
@@ -1,3 +1,24 @@
+// @flow
+type LoaderConfig = {
+  loader?: string,
+  options?: Object
+}
+
+type LoaderConfigFactory = (id: ?string, defaultConfig: LoaderConfig) => LoaderConfig
+
+type RuleConfig = {
+  test?: RegExp,
+  include?: RegExp,
+  exclude?: RegExp,
+  loader?: string,
+  options?: Object,
+  use?: UseConfig
+}
+
+type RuleConfigFactory = (?string, RuleConfig) => ?RuleConfig
+
+type UseConfig = Array<string | LoaderConfig>
+
 import path from 'path'
 
 import autoprefixer from 'autoprefixer'
@@ -22,9 +43,13 @@ import StatusPlugin from './WebpackStatusPlugin'
 let replaceArrayMerge = merge({customizeArray(a, b, key) { return b }})
 
 /**
- * Merge webpack rule config ({test, loader+options|use, include, exclude, ...}) objects.
+ * Merge webpack rule config objects.
  */
-export function mergeRuleConfig(defaultConfig = {}, buildConfig = {}, userConfig = {}) {
+export function mergeRuleConfig(
+  defaultConfig: RuleConfig,
+  buildConfig: RuleConfig = {},
+  userConfig: RuleConfig = {}
+): RuleConfig {
   let rule
   // Omit the default loader and options if the user is configuring their own
   if (defaultConfig.loader && (userConfig.loader || userConfig.use)) {
@@ -44,11 +69,15 @@ export function mergeRuleConfig(defaultConfig = {}, buildConfig = {}, userConfig
 }
 
 /**
- * Merge webpack loader config ({loader, options}) objects.
+ * Merge webpack loader config objects.
  */
-export function mergeLoaderConfig(defaultConfig = {}, buildConfig = {}, userConfig = {}) {
+export function mergeLoaderConfig(
+  defaultConfig: LoaderConfig,
+  buildConfig: LoaderConfig = {},
+  userConfig: LoaderConfig = {}
+): RuleConfig {
   let loader
-  // If the loader s being changed, only use the provided config
+  // If the loader is being changed, only use the provided config
   if (userConfig.loader) {
     loader = {...userConfig}
   }
@@ -65,8 +94,11 @@ export function mergeLoaderConfig(defaultConfig = {}, buildConfig = {}, userConf
  * Create a function which configures a rule identified by a unique id, with
  * the option to override defaults with build-specific and user config.
  */
-export let ruleConfigFactory = (buildConfig, userConfig = {}) =>
-  (id, defaultConfig) => {
+export function createRuleConfigFactory(
+  buildConfig: {[key: string]: RuleConfig} = {},
+  userConfig: {[key: string]: RuleConfig} = {}
+): RuleConfigFactory {
+  return function(id: ?string, defaultConfig: RuleConfig): ?RuleConfig {
     if (id) {
       // Allow the user to turn off rules by configuring them with false
       if (userConfig[id] === false) {
@@ -77,32 +109,37 @@ export let ruleConfigFactory = (buildConfig, userConfig = {}) =>
     }
     return defaultConfig
   }
+}
 
 /**
  * Create a function which configures a loader identified by a unique id, with
  * the option to override defaults with build-specific and user config.
  */
-export let loaderConfigFactory = (buildConfig, userConfig = {}) =>
-  (id, defaultConfig) => {
+export function createLoaderConfigFactory(
+  buildConfig: {[key: string]: LoaderConfig} = {},
+  userConfig: {[key: string]: LoaderConfig} = {}
+): LoaderConfigFactory {
+  return function (id: ?string, defaultConfig: LoaderConfig): LoaderConfig {
     if (id) {
       let loader = mergeLoaderConfig(defaultConfig, buildConfig[id], userConfig[id])
       return loader
     }
     return defaultConfig
   }
+}
 
 /**
- * Create a function which applies a prefix to a given name when a prefix is
- * given, unless the prefix ends with a name, in which case the prefix itself is
+ * Create a function which applies a prefix to a name when a prefix is given,
+ * unless the prefix ends with the name, in which case the prefix itself is
  * returned.
  * The latter rule is to allow rules created for CSS preprocessor plugins to
  * be given unique ids for user configuration without duplicating the name of
  * the rule.
- * e.g.: styleRuleName('sass')('css') => 'sass-css'
- *       styleRuleName('sass')('sass') => 'sass' (as opposed to 'sass-sass')
+ * e.g.: loaderConfigName('sass')('css') => 'sass-css'
+ *       loaderConfigName('sass')('sass') => 'sass' (as opposed to 'sass-sass')
  */
-export let styleRuleName = (prefix) =>
-  (name) => {
+export let loaderConfigName = (prefix: ?string) =>
+  (name: string): string => {
     if (prefix && prefix.endsWith(name)) {
       return prefix
     }
@@ -110,26 +147,36 @@ export let styleRuleName = (prefix) =>
   }
 
 /**
- * Create a default style-handling pipeline for either a static build (default)
- * or a server build.
+ * Create a list of chained loader config objects for a static build (default)
+ * or serving.
  */
-export function createStyleLoaders(loader, server, userWebpackConfig, {
-  preprocessor = null,
-  prefix = null,
-} = {}) {
-  let name = styleRuleName(prefix)
-  let styleLoader = loader(name('style'), {
+export function createStyleLoaders(
+  createLoader: (?string, LoaderConfig) => LoaderConfig,
+  userWebpackConfig: Object,
+  options: {
+    preprocessor?: ?Object,
+    prefix?: ?string,
+    server?: ?boolean
+  } = {},
+): UseConfig {
+  let {
+    preprocessor = null,
+    prefix = null,
+    server = false,
+  } = options
+  let name = loaderConfigName(prefix)
+  let styleLoader = createLoader(name('style'), {
     loader: require.resolve('style-loader'),
   })
   let loaders = [
-    loader(name('css'), {
+    createLoader(name('css'), {
       loader: require.resolve('css-loader'),
       options: {
         // Apply postcss-loader to @imports
         importLoaders: 1,
       },
     }),
-    loader(name('postcss'), {
+    createLoader(name('postcss'), {
       loader: require.resolve('postcss-loader'),
       options: {
         plugins: createDefaultPostCSSPlugins(userWebpackConfig),
@@ -138,7 +185,10 @@ export function createStyleLoaders(loader, server, userWebpackConfig, {
   ]
 
   if (preprocessor) {
-    loaders.push(loader(name(preprocessor.id), preprocessor.config))
+    loaders.push(createLoader(
+      preprocessor.id ? name(preprocessor.id) : null,
+      preprocessor.config
+    ))
   }
 
   if (server) {
@@ -154,6 +204,131 @@ export function createStyleLoaders(loader, server, userWebpackConfig, {
 }
 
 /**
+ * Create style rules for nwb >= 0.16.
+ */
+function createStyleRules(
+  server: boolean,
+  userWebpackConfig: Object,
+  pluginConfig: Object,
+  createRule: RuleConfigFactory,
+  createLoader: LoaderConfigFactory
+): Array<?RuleConfig> {
+  let styleConfig = userWebpackConfig.styles || {}
+  let styleRules = []
+
+  // Configured styles rules, with individual loader configuration as part of
+  // the definition.
+  Object.keys(styleConfig).forEach(type => {
+    let test, preprocessor
+    if (type === 'css') {
+      test = /\.css$/
+    }
+    else {
+      let preprocessorConfig = pluginConfig.cssPreprocessors[type]
+      test = preprocessorConfig.test
+      preprocessor = {id: null, config: {loader: preprocessorConfig.loader}}
+    }
+    let ruleConfigs = [].concat(...styleConfig[type])
+    ruleConfigs.forEach(ruleConfig => {
+      let {loaders: loaderConfig, ...topLevelRuleConfig} = ruleConfig
+      // Empty build config, as all loader config for custom style rules will be
+      // provided by the user.
+      let styleRuleLoader = createLoaderConfigFactory({}, loaderConfig)
+      styleRules.push({
+        test,
+        use: createStyleLoaders(styleRuleLoader, userWebpackConfig, {preprocessor, server}),
+        ...topLevelRuleConfig,
+      })
+    })
+  })
+
+  // Default CSS rule when nothing is configured, tweakable via webpack.rules by
+  // unique id.
+  if (!('css' in styleConfig)) {
+    styleRules.push(
+      createRule('css-rule', {
+        test: /\.css$/,
+        use: createStyleLoaders(createLoader, userWebpackConfig, {server}),
+      })
+    )
+  }
+
+  // Default rule for each CSS preprocessor plugin when nothing is configured,
+  // tweakable via webpack.rules by unique id.
+  if (pluginConfig.cssPreprocessors) {
+    Object.keys(pluginConfig.cssPreprocessors).forEach(id => {
+      if (id in styleConfig) return
+      let {test, loader: preprocessorLoader} = pluginConfig.cssPreprocessors[id]
+      styleRules.push(
+        createRule(`${id}-rule`, {
+          test,
+          use: createStyleLoaders(createLoader, userWebpackConfig, {
+            prefix: id,
+            preprocessor: {id, config: {loader: preprocessorLoader}},
+            server,
+          })
+        })
+      )
+    })
+  }
+
+  return styleRules
+}
+
+// TODO Remove in a future version
+/**
+ * Create default style rules for nwb < 0.16.
+ */
+function createLegacyStyleRules(
+  server: boolean,
+  userWebpackConfig: Object,
+  pluginConfig: Object,
+  createRule: RuleConfigFactory,
+  createLoader: LoaderConfigFactory
+): Array<?RuleConfig> {
+  let styleRules = [
+    createRule('css-pipeline', {
+      test: /\.css$/,
+      use: createStyleLoaders(createLoader, userWebpackConfig, {server}),
+      exclude: /node_modules/,
+    }),
+    createRule('vendor-css-pipeline', {
+      test: /\.css$/,
+      use: createStyleLoaders(createLoader, userWebpackConfig, {prefix: 'vendor', server}),
+      include: /node_modules/,
+    })
+  ]
+
+  if (pluginConfig.cssPreprocessors) {
+    Object.keys(pluginConfig.cssPreprocessors).forEach(id => {
+      let {test, loader: preprocessorLoader} = pluginConfig.cssPreprocessors[id]
+      styleRules.push(
+        createRule(`${id}-pipeline`, {
+          test,
+          use: createStyleLoaders(createLoader, userWebpackConfig, {
+            prefix: id,
+            preprocessor: {id, config: {loader: preprocessorLoader}},
+            server,
+          }),
+          exclude: /node_modules/
+        }),
+        createRule(`vendor-${id}-pipeline`, {
+          test,
+          use: createStyleLoaders(createLoader, userWebpackConfig, {
+            prefix: `vendor-${id}`,
+            preprocessor: {id, config: {loader: preprocessorLoader}},
+            server,
+          }),
+          include: /node_modules/
+        })
+      )
+    })
+  }
+
+  return styleRules
+}
+
+/**
  * Final webpack rules config consists of:
  * - the default set of rules created in this function, with build and user
  *   config tweaks based on rule id.
@@ -163,9 +338,14 @@ export function createStyleLoaders(loader, server, userWebpackConfig, {
  *   tweaks based on loader id.
  * - extra rules defined in user config.
  */
-export function createRules(server, buildConfig = {}, userWebpackConfig = {}, pluginConfig = {}) {
-  let rule = ruleConfigFactory(buildConfig, userWebpackConfig.rules)
-  let loader = loaderConfigFactory(buildConfig, userWebpackConfig.rules)
+export function createRules(
+  server: boolean,
+  buildConfig: Object = {},
+  userWebpackConfig: Object = {},
+  pluginConfig: Object = {}
+) {
+  let createRule = createRuleConfigFactory(buildConfig, userWebpackConfig.rules)
+  let createLoader = createLoaderConfigFactory(buildConfig, userWebpackConfig.rules)
 
   // Default options for url-loader
   let urlLoaderOptions = {
@@ -176,7 +356,7 @@ export function createRules(server, buildConfig = {}, userWebpackConfig = {}, pl
   }
 
   let rules = [
-    rule('babel', {
+    createRule('babel', {
       test: /\.js$/,
       loader: require.resolve('babel-loader'),
       exclude: process.env.NWB_TEST ? /(node_modules|nwb[\\/]polyfills\.js$)/ : /node_modules/,
@@ -187,91 +367,69 @@ export function createRules(server, buildConfig = {}, userWebpackConfig = {}, pl
         cacheDirectory: true,
       }
     }),
-    rule('css-pipeline', {
-      test: /\.css$/,
-      use: createStyleLoaders(loader, server, userWebpackConfig),
-      exclude: /node_modules/,
-    }),
-    rule('vendor-css-pipeline', {
-      test: /\.css$/,
-      use: createStyleLoaders(loader, server, userWebpackConfig, {
-        prefix: 'vendor',
-      }),
-      include: /node_modules/,
-    }),
-    rule('graphics', {
+    createRule('graphics', {
       test: /\.(gif|png|webp)$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
-    rule('svg', {
+    createRule('svg', {
       test: /\.svg$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
-    rule('jpeg', {
+    createRule('jpeg', {
       test: /\.jpe?g$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
-    rule('fonts', {
+    createRule('fonts', {
       test: /\.(eot|otf|ttf|woff|woff2)$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
-    rule('video', {
+    createRule('video', {
       test: /\.(mp4|ogg|webm)$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
-    rule('audio', {
-      test: /\.(wav|mp3|m4a|aac|oga)(\?.*)?$/,
+    createRule('audio', {
+      test: /\.(wav|mp3|m4a|aac|oga)$/,
       loader: require.resolve('url-loader'),
       options: {...urlLoaderOptions},
     }),
     // Extra rules from build config, still configurable via user config when
     // the rules specify an id.
     ...createExtraRules(buildConfig.extra, userWebpackConfig.rules),
-  ].filter(rule => rule != null)
+  ]
 
-  if (pluginConfig.cssPreprocessors) {
-    Object.keys(pluginConfig.cssPreprocessors).forEach(id => {
-      let {test, loader: preprocessorLoader} = pluginConfig.cssPreprocessors[id]
-      rules.push(
-        rule(`${id}-pipeline`, {
-          test,
-          use: createStyleLoaders(loader, server, userWebpackConfig, {
-            prefix: id,
-            preprocessor: {id, config: {loader: preprocessorLoader}},
-          }),
-          exclude: /node_modules/
-        })
-      )
-      rules.push(
-        rule(`vendor-${id}-pipeline`, {
-          test,
-          use: createStyleLoaders(loader, server, userWebpackConfig, {
-            prefix: `vendor-${id}`,
-            preprocessor: {id, config: {loader: preprocessorLoader}},
-          }),
-          include: /node_modules/
-        })
-      )
-    })
+  // Add rules with chained style loaders, using ExtractTextPlugin for builds
+  if (userWebpackConfig.styles === 'old') {
+    // nwb <= 0.15 default
+    rules = rules.concat(createLegacyStyleRules(
+      server, userWebpackConfig, pluginConfig, createRule, createLoader
+    ))
+  }
+  else if (userWebpackConfig.styles !== false) {
+    rules = rules.concat(createStyleRules(
+      server, userWebpackConfig, pluginConfig, createRule, createLoader
+    ))
   }
 
-  return rules
+  return rules.filter(rule => rule != null)
 }
 
 /**
  * Create rules from rule definitions which may include an id attribute for
  * user customisation. It's assumed these are being created from build config.
  */
-export function createExtraRules(extraRules = [], userConfig = {}) {
-  let rule = ruleConfigFactory({}, userConfig)
+export function createExtraRules(
+  extraRules: Object[] = [],
+  userConfig: Object = {}
+): Array<?RuleConfig> {
+  let createRule = createRuleConfigFactory({}, userConfig)
   return extraRules.map(extraRule => {
-    let {id, ...ruleConfig} = extraRules
-    return rule(id, ruleConfig)
+    let {id, ...ruleConfig} = extraRule
+    return createRule(id, ruleConfig)
   })
 }
 
@@ -327,7 +485,11 @@ function getCopyPluginArgs(buildConfig, userConfig) {
  * - any extra plugins defined in build and user config (extra user plugins are
  *   not handled here, but by the final merge of webpack.extra config).
  */
-export function createPlugins(server, buildConfig = {}, userConfig = {}) {
+export function createPlugins(
+  server: boolean,
+  buildConfig: Object = {},
+  userConfig: Object = {}
+) {
   let development = process.env.NODE_ENV === 'development'
   let production = process.env.NODE_ENV === 'production'
 
@@ -361,6 +523,7 @@ export function createPlugins(server, buildConfig = {}, userConfig = {}) {
     // Extract CSS required as modules out into files
     let cssFilename = production ? `[name].[contenthash:8].css` : '[name].css'
     plugins.push(new ExtractTextPlugin({
+      allChunks: true,
       filename: cssFilename,
       ...userConfig.extractText,
     }))
@@ -481,12 +644,12 @@ export const COMPAT_CONFIGS = {
       'react/lib/ReactContext': true,
     }
   },
-  moment({locales}) {
+  moment(options: {locales: string[]}) {
     return {
       plugins: [
         new webpack.ContextReplacementPlugin(
           /moment[/\\]locale$/,
-          new RegExp(`^\\.\\/(${locales.join('|')})$`)
+          new RegExp(`^\\.\\/(${options.locales.join('|')})$`)
         )
       ]
     }
@@ -509,7 +672,7 @@ export const COMPAT_CONFIGS = {
  * config.
  * Returns null if there's nothing to merge based on user config.
  */
-export function getCompatConfig(userCompatConfig = {}) {
+export function getCompatConfig(userCompatConfig: Object = {}): ?Object {
   let configs = []
   Object.keys(userCompatConfig).map(lib => {
     if (!userCompatConfig[lib]) return
@@ -542,7 +705,11 @@ function addPolyfillsToEntry(entry) {
  * Create a webpack config with a curated set of default rules suitable for
  * creating a static build (default) or serving an app with hot reloading.
  */
-export default function createWebpackConfig(buildConfig, pluginConfig = {}, userConfig = {}) {
+export default function createWebpackConfig(
+  buildConfig: Object,
+  pluginConfig: Object = {},
+  userConfig: Object = {}
+): Object {
   debug('createWebpackConfig buildConfig: %s', deepToString(buildConfig))
 
   // Final webpack config is primarily driven by build configuration for the nwb